mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.51.0",
    "boto3>=1.38.19",
    "mcp>=1.9.0",
    "openai>=1.79.0",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
]
